Changed script to be idempotent

+++ IaaS2PaaSWeb/Environments/CustomScripts/vsDevWorkstationConfig.ps1
@@ -13,23 +13,26 @@ Param (
 
 $current_dir = $PSScriptRoot
 if (!$current_dir) { $current_dir = Get-Location }
-#Stop on any error
-$ErrorActionPreference = "Stop"
 
 ## Clone Repo
 $repo_path = 'c:\Source'
 if (!(Test-Path $repo_path)) { 
-    New-Item -Path $repo_path 
+    mkdir $repo_path 
     cd $repo_path
     if (!(Test-Path "$repo_path\AppWorkshop")) {
         git clone $repoUri
     }
+	cd $current_dir 
 }
 
 ## Install Chocolatey and packages
-Invoke-Expression ((New-Object Net.WebClient).DownloadString('https://chocolatey.org/install.ps1')) 
+if ((Get-Command choco -ErrorAction SilentlyContinue) -eq $null) {
+	Invoke-Expression ((New-Object Net.WebClient).DownloadString('https://chocolatey.org/install.ps1')) 
+}
 #Install NuGet
-& choco install nuget.commandline -y
+if ((Get-Command nuget -ErrorAction SilentlyContinue) -eq $null) {
+	& choco install nuget.commandline -y
+}
 
 ## Build and Package App
 
@@ -44,7 +47,7 @@ if (!(Test-Path $build_bat_file)) {
     Add-Content -Path $build_bat_file -Value "call `"C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\Common7\Tools\vsdevcmd\core\msbuild.bat`""
     Add-Content -Path $build_bat_file -Value "msbuild `"C:\Source\AppWorkshop\IaaS2PaaSWeb\PartsUnlimitedWebsite\PartsUnlimitedWebsite.csproj`" /p:DeployOnBuild=true /p:WebPublishMethod=Package /p:PackageAsSingleFile=true /p:SkipInvalidConfigurations=true"
 }
-cmd.exe /c $build_bat_file > build.log
+cmd.exe /c $build_bat_file
 
 ## Deploy Webapp
 
@@ -62,4 +65,6 @@ C:\Source\AppWorkshop\IaaS2PaaSWeb\PartsUnlimitedWebsite\obj\Debug\Package\parts
 ## Add startup bat to install additional packages on sign in
 $choco_exe = "C:\ProgramData\chocolatey\bin\choco.exe"
 $install_packages_bat = "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\StartUp\install_packages.bat"
-Set-Content -Path $install_packages_bat -Value "$choco_exe install postman googlechrome -y"
+if (!(Test-Path $install_packages_bat)) {
+	Set-Content -Path $install_packages_bat -Value "$choco_exe install postman googlechrome -y"
+}
